separate jupyter tests
